src/third_party/blink/web_tests/virtual/web-components-v0-disabled/fast/dom/custom/element-type-expected.txt

CONSOLE ERROR: line 28: Uncaught TypeError: document.registerElement is not a function
Test that changing an instantiated custom element's 'is' attribute does not affect its element type.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS window.testRunner is non-null.
FAIL successfullyParsed should be true. Was false.

TEST COMPLETE
